// libs/simdata-api/nest-client/src/lib/model/hdf5-dataset.ts
/**
 * simdata-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HDF5Attribute } from './hdf5-attribute';

export interface HDF5Dataset {
  name: string;
  shape: Array<number>;
  attributes: Array<HDF5Attribute>;
}